refac(virtualCompletion): into one function
